Move states and transitions to public
