+++ .python-version
@@ -1,1 +1,1 @@
-3.7
+3.8
