+++ src/pyfoma/transition.py
